# nfvo_server/nfvo_server/models/assignment.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from nfvo_server.models.base_model_ import Model
from nfvo_server import util


class Assignment(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, sfcr_id: int=None, vnf_instace_ids: List[int]=None):  # noqa: E501
        """Assignment - a model defined in Swagger

        :param sfcr_id: The sfcr_id of this Assignment.  # noqa: E501
        :type sfcr_id: int
        :param vnf_instace_ids: The vnf_instace_ids of this Assignment.  # noqa: E501
        :type vnf_instace_ids: List[int]
        """
        self.swagger_types = {
            'sfcr_id': int,
            'vnf_instace_ids': List[int]
        }

        self.attribute_map = {
            'sfcr_id': 'sfcr_id',
            'vnf_instace_ids': 'vnf_instace_ids'
        }

        self._sfcr_id = sfcr_id
        self._vnf_instace_ids = vnf_instace_ids

    @classmethod
    def from_dict(cls, dikt) -> 'Assignment':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Assignment of this Assignment.  # noqa: E501
        :rtype: Assignment
        """
        return util.deserialize_model(dikt, cls)

    @property
    def sfcr_id(self) -> int:
        """Gets the sfcr_id of this Assignment.


        :return: The sfcr_id of this Assignment.
        :rtype: int
        """
        return self._sfcr_id

    @sfcr_id.setter
    def sfcr_id(self, sfcr_id: int):
        """Sets the sfcr_id of this Assignment.


        :param sfcr_id: The sfcr_id of this Assignment.
        :type sfcr_id: int
        """

        self._sfcr_id = sfcr_id

    @property
    def vnf_instace_ids(self) -> List[int]:
        """Gets the vnf_instace_ids of this Assignment.


        :return: The vnf_instace_ids of this Assignment.
        :rtype: List[int]
        """
        return self._vnf_instace_ids

    @vnf_instace_ids.setter
    def vnf_instace_ids(self, vnf_instace_ids: List[int]):
        """Sets the vnf_instace_ids of this Assignment.


        :param vnf_instace_ids: The vnf_instace_ids of this Assignment.
        :type vnf_instace_ids: List[int]
        """

        self._vnf_instace_ids = vnf_instace_ids