# nfvo_server/nfvo_server/models/route.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from nfvo_server.models.base_model_ import Model
from nfvo_server.models.route_hops import RouteHops  # noqa: F401,E501
from nfvo_server import util


class Route(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, sfcr_id: int=None, hops: List[RouteHops]=None):  # noqa: E501
        """Route - a model defined in Swagger

        :param sfcr_id: The sfcr_id of this Route.  # noqa: E501
        :type sfcr_id: int
        :param hops: The hops of this Route.  # noqa: E501
        :type hops: List[RouteHops]
        """
        self.swagger_types = {
            'sfcr_id': int,
            'hops': List[RouteHops]
        }

        self.attribute_map = {
            'sfcr_id': 'sfcr_id',
            'hops': 'hops'
        }

        self._sfcr_id = sfcr_id
        self._hops = hops

    @classmethod
    def from_dict(cls, dikt) -> 'Route':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Route of this Route.  # noqa: E501
        :rtype: Route
        """
        return util.deserialize_model(dikt, cls)

    @property
    def sfcr_id(self) -> int:
        """Gets the sfcr_id of this Route.


        :return: The sfcr_id of this Route.
        :rtype: int
        """
        return self._sfcr_id

    @sfcr_id.setter
    def sfcr_id(self, sfcr_id: int):
        """Sets the sfcr_id of this Route.


        :param sfcr_id: The sfcr_id of this Route.
        :type sfcr_id: int
        """

        self._sfcr_id = sfcr_id

    @property
    def hops(self) -> List[RouteHops]:
        """Gets the hops of this Route.


        :return: The hops of this Route.
        :rtype: List[RouteHops]
        """
        return self._hops

    @hops.setter
    def hops(self, hops: List[RouteHops]):
        """Sets the hops of this Route.


        :param hops: The hops of this Route.
        :type hops: List[RouteHops]
        """

        self._hops = hops