# nfvo_server/nfvo_server/models/topology.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from nfvo_server.models.base_model_ import Model
from nfvo_server import util


class Topology(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, node_list: List[int]=None, edge_list: List[int]=None):  # noqa: E501
        """Topology - a model defined in Swagger

        :param node_list: The node_list of this Topology.  # noqa: E501
        :type node_list: List[int]
        :param edge_list: The edge_list of this Topology.  # noqa: E501
        :type edge_list: List[int]
        """
        self.swagger_types = {
            'node_list': List[int],
            'edge_list': List[int]
        }

        self.attribute_map = {
            'node_list': 'node_list',
            'edge_list': 'edge_list'
        }

        self._node_list = node_list
        self._edge_list = edge_list

    @classmethod
    def from_dict(cls, dikt) -> 'Topology':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Topology of this Topology.  # noqa: E501
        :rtype: Topology
        """
        return util.deserialize_model(dikt, cls)

    @property
    def node_list(self) -> List[int]:
        """Gets the node_list of this Topology.


        :return: The node_list of this Topology.
        :rtype: List[int]
        """
        return self._node_list

    @node_list.setter
    def node_list(self, node_list: List[int]):
        """Sets the node_list of this Topology.


        :param node_list: The node_list of this Topology.
        :type node_list: List[int]
        """

        self._node_list = node_list

    @property
    def edge_list(self) -> List[int]:
        """Gets the edge_list of this Topology.


        :return: The edge_list of this Topology.
        :rtype: List[int]
        """
        return self._edge_list

    @edge_list.setter
    def edge_list(self, edge_list: List[int]):
        """Sets the edge_list of this Topology.


        :param edge_list: The edge_list of this Topology.
        :type edge_list: List[int]
        """

        self._edge_list = edge_list